Support reusing existing storage containers across task providers (#687)
* Support for existing storage containers in aws and gcp. (#651)

* Update config schema.

* AWS support for existing S3 buckets.

* Existing bucket support for gcp.

* Add mocks and tests to existing bucket resource in aws.

* Update docs with new pre-allocated container fields.

* Support using pre-allocated blob containers in azure. (#660)

* AWS support for existing S3 buckets.

* Existing bucket support for gcp.

* Fix subdirectory in rclone remote.

* Blob container generates the rclone connection string.

* Introduce a type for generating rclone connection strings.

* Azure support for reusable blob containers.

* Update docs.

* Fix path prefix.

* Initialize s3 existing bucket with RemoteStorage struct.

* Update gcp and aws existing bucket data sources to align with the azure data source.

Use common.RemoteStorage to initialize the data sources.
Using rclone to verify storage during Read.
Remove aws s3 client mocks and tests that rely on them.

* Fix comment.

* K8s support for specifying an existing persistent volume claim (#661)

* K8s support for specifying an existing persistent volume claim.


* Update use of Identifier struct.

* Combine container and container_path config keys.

* Update docs.

* Use split function from rclone.

+++ task/az/task.go
@@ -40,23 +40,36 @@ func New(ctx context.Context, cloud common.Cloud, identifier common.Identifier, 
 		t.Client,
 		t.Identifier,
 	)
-	t.Resources.StorageAccount = resources.NewStorageAccount(
-		t.Client,
-		t.Identifier,
-		t.Resources.ResourceGroup,
-	)
-	t.Resources.BlobContainer = resources.NewBlobContainer(
-		t.Client,
-		t.Identifier,
-		t.Resources.ResourceGroup,
-		t.Resources.StorageAccount,
-	)
+	var bucketCredentials common.StorageCredentials
+	if task.RemoteStorage != nil {
+		// If a subdirectory was not specified, the task id will
+		// be used.
+		if task.RemoteStorage.Path == "" {
+			task.RemoteStorage.Path = t.Identifier.Short()
+		}
+		bucket := resources.NewExistingBlobContainer(t.Client, *task.RemoteStorage)
+		t.DataSources.BlobContainer = bucket
+		bucketCredentials = bucket
+	} else {
+		t.Resources.StorageAccount = resources.NewStorageAccount(
+			t.Client,
+			t.Identifier,
+			t.Resources.ResourceGroup,
+		)
+		blobContainer := resources.NewBlobContainer(
+			t.Client,
+			t.Identifier,
+			t.Resources.ResourceGroup,
+			t.Resources.StorageAccount,
+		)
+		t.Resources.BlobContainer = blobContainer
+		bucketCredentials = blobContainer
+	}
 	t.DataSources.Credentials = resources.NewCredentials(
 		t.Client,
 		t.Identifier,
 		t.Resources.ResourceGroup,
-		t.Resources.StorageAccount,
-		t.Resources.BlobContainer,
+		bucketCredentials,
 	)
 	t.Resources.VirtualNetwork = resources.NewVirtualNetwork(
 		t.Client,
@@ -94,17 +107,18 @@ type Task struct {
 	Identifier  common.Identifier
 	Attributes  common.Task
 	DataSources struct {
-		*resources.Credentials
-		*resources.PermissionSet
+		Credentials   *resources.Credentials
+		PermissionSet *resources.PermissionSet
+		BlobContainer *resources.ExistingBlobContainer
 	}
 	Resources struct {
-		*resources.ResourceGroup
-		*resources.StorageAccount
-		*resources.BlobContainer
-		*resources.VirtualNetwork
-		*resources.Subnet
-		*resources.SecurityGroup
-		*resources.VirtualMachineScaleSet
+		ResourceGroup          *resources.ResourceGroup
+		StorageAccount         *resources.StorageAccount
+		BlobContainer          *resources.BlobContainer
+		VirtualNetwork         *resources.VirtualNetwork
+		Subnet                 *resources.Subnet
+		SecurityGroup          *resources.SecurityGroup
+		VirtualMachineScaleSet *resources.VirtualMachineScaleSet
 	}
 }
 
@@ -113,13 +127,23 @@ func (t *Task) Create(ctx context.Context) error {
 	steps := []common.Step{{
 		Description: "Creating ResourceGroup...",
 		Action:      t.Resources.ResourceGroup.Create,
-	}, {
-		Description: "Creating StorageAccount...",
-		Action:      t.Resources.StorageAccount.Create,
-	}, {
-		Description: "Creating BlobContainer...",
-		Action:      t.Resources.BlobContainer.Create,
-	}, {
+	}}
+	if t.Resources.BlobContainer != nil {
+		steps = append(steps, []common.Step{{
+			Description: "Creating StorageAccount...",
+			Action:      t.Resources.StorageAccount.Create,
+		}, {
+			Description: "Creating BlobContainer...",
+			Action:      t.Resources.BlobContainer.Create,
+		}}...)
+	} else if t.DataSources.BlobContainer != nil {
+		steps = append(steps, common.Step{
+			Description: "Reading BlobContainer...",
+			Action:      t.DataSources.BlobContainer.Read,
+		})
+	}
+
+	steps = append(steps, []common.Step{{
 		Description: "Creating Credentials...",
 		Action:      t.DataSources.Credentials.Read,
 	}, {
@@ -134,7 +158,7 @@ func (t *Task) Create(ctx context.Context) error {
 	}, {
 		Description: "Creating VirtualMachineScaleSet...",
 		Action:      t.Resources.VirtualMachineScaleSet.Create,
-	}}
+	}}...)
 	if t.Attributes.Environment.Directory != "" {
 		steps = append(steps, common.Step{
 			Description: "Uploading Directory...",
@@ -160,13 +184,23 @@ func (t *Task) Read(ctx context.Context) error {
 	steps := []common.Step{{
 		Description: "Reading ResourceGroup...",
 		Action:      t.Resources.ResourceGroup.Read,
-	}, {
-		Description: "Reading StorageAccount...",
-		Action:      t.Resources.StorageAccount.Read,
-	}, {
-		Description: "Reading BlobContainer...",
-		Action:      t.Resources.BlobContainer.Read,
-	}, {
+	}}
+	if t.Resources.BlobContainer != nil {
+		steps = append(steps, []common.Step{{
+			Description: "Reading StorageAccount...",
+			Action:      t.Resources.StorageAccount.Read,
+		}, {
+			Description: "Reading BlobContainer...",
+			Action:      t.Resources.BlobContainer.Read,
+		}}...)
+	} else {
+		steps = append(steps, common.Step{
+			Description: "Reading BlobContainer...",
+			Action:      t.DataSources.BlobContainer.Read,
+		})
+	}
+
+	steps = append(steps, []common.Step{{
 		Description: "Reading Credentials...",
 		Action:      t.DataSources.Credentials.Read,
 	}, {
@@ -181,7 +215,7 @@ func (t *Task) Read(ctx context.Context) error {
 	}, {
 		Description: "Reading VirtualMachineScaleSet...",
 		Action:      t.Resources.VirtualMachineScaleSet.Read,
-	}}
+	}}...)
 	if err := common.RunSteps(ctx, steps); err != nil {
 		return err
 	}
@@ -207,7 +241,10 @@ func (t *Task) Delete(ctx context.Context) error {
 					}
 					return nil
 				},
-			}, {
+			}}
+		}
+		if t.Resources.BlobContainer != nil {
+			steps = append(steps, common.Step{
 				Description: "Emptying Bucket...",
 				Action: func(ctx context.Context) error {
 					err := machine.Delete(ctx, t.DataSources.Credentials.Resource["RCLONE_REMOTE"])
@@ -216,7 +253,7 @@ func (t *Task) Delete(ctx context.Context) error {
 					}
 					return nil
 				},
-			}}
+			})
 		}
 	}
 	steps = append(steps, []common.Step{{
@@ -231,16 +268,21 @@ func (t *Task) Delete(ctx context.Context) error {
 	}, {
 		Description: "Deleting VirtualNetwork...",
 		Action:      t.Resources.VirtualNetwork.Delete,
-	}, {
-		Description: "Deleting BlobContainer...",
-		Action:      t.Resources.BlobContainer.Delete,
-	}, {
-		Description: "Deleting StorageAccount...",
-		Action:      t.Resources.StorageAccount.Delete,
-	}, {
+	}}...)
+
+	if t.Resources.BlobContainer != nil {
+		steps = append(steps, []common.Step{{
+			Description: "Deleting BlobContainer...",
+			Action:      t.Resources.BlobContainer.Delete,
+		}, {
+			Description: "Deleting StorageAccount...",
+			Action:      t.Resources.StorageAccount.Delete,
+		}}...)
+	}
+	steps = append(steps, common.Step{
 		Description: "Deleting ResourceGroup...",
 		Action:      t.Resources.ResourceGroup.Delete,
-	}}...)
+	})
 	if err := common.RunSteps(ctx, steps); err != nil {
 		return err
 	}
